+++ src/parsers/hsl.ts
@@ -7,7 +7,8 @@ import { isHSLObject } from "../utils/typeGuards";
  * hsl[a]( <number | percentage | angle> <number | percentage> <number | percentage> [ / <alpha-value> ]? )
  * @see https://www.w3.org/TR/css-color-4/#the-hsl-notation
  */
-const HSLA_RE = /hsla?\s*\(\s*([+-]?\d*\.?\d+%?),?\s*(\d*\.?\d+%?),?\s*(\d*\.?\d+%?),?\s*\/?\s*?(\d*\.?\d+%?)?\s*\)/gi;
+const HSLA_RE =
+  /hsla?\s*\(\s*([+-]?\d*\.?\d+%?)\s*,?\s*(\d*\.?\d+%?)\s*,?\s*(\d*\.?\d+%?)\s*,?\s*\/?\s*?(\d*\.?\d+%?)?\s*\)/gi;
 
 function createReturnArr({ h, s, l, a }: Ihsla): [Irgba, TFormat] {
   return [HSLtoRGB({ h: adjustHue(h), s: clamp(0, s, 100), l: clamp(0, l, 100), a: adjustAlpha(a) }), "hsl"];