+++ src/parsers/xyz.ts
@@ -5,12 +5,13 @@ import { adjustAlpha, clamp } from "../utils/numeric";
 import { isXYZObject } from "../utils/typeGuards";
 
 /**
- * color(xyz[a] <number | percentage>{3} [ / <alpha-value> ]? )
+ * color(xyz[a] <number>{3} [ / <alpha-value> ]? )
  * @note Percentages are not allowed (for non alpha channels)
  * @see https://en.wikipedia.org/wiki/CIE_1931_color_space
  * @see https://www.w3.org/TR/css-color-4/#predefined
  */
-const XYZA_RE = /color\s*\(\s*xyza?\s*(\d*\.?\d+),?\s*(\d*\.?\d+),?\s*(\d*\.?\d+),?\s*\/?\s*?(\d*\.?\d+%?)?\s*\)/gi;
+const XYZA_RE =
+  /color\s*\(\s*xyza?\s*(\d*\.?\d+)\s*,?\s*(\d*\.?\d+)\s*,?\s*(\d*\.?\d+)\s*,?\s*\/?\s*?(\d*\.?\d+%?)?\s*\)/gi;
 
 function createReturnArr({ x, y, z, a }: Ixyza): [Irgba, TFormat] {
   return [